Moved count_down_latch classes into own files.
