+++ lib/concurrent/atomic/count_down_latch.rb
@@ -1,6 +1,65 @@
-require 'concurrent/synchronization'
+require 'concurrent/atomic/mutex_count_down_latch'
+require 'concurrent/atomic/java_count_down_latch'
+require 'concurrent/utility/engine'
 
 module Concurrent
+
+  ###################################################################
+
+  # @!macro [new] count_down_latch_method_initialize
+  #
+  #   Create a new `CountDownLatch` with the initial `count`.
+  #
+  #   @param [new] count the initial count
+  #
+  #   @raise [ArgumentError] if `count` is not an integer or is less than zero
+
+  # @!macro [new] count_down_latch_method_wait
+  #
+  #   Block on the latch until the counter reaches zero or until `timeout` is reached.
+  #
+  #   @param [Fixnum] timeout the number of seconds to wait for the counter or `nil`
+  #     to block indefinitely
+  #   @return [Boolean] `true` if the `count` reaches zero else false on `timeout`
+
+  # @!macro [new] count_down_latch_method_count_down
+  #
+  #   Signal the latch to decrement the counter. Will signal all blocked threads when
+  #   the `count` reaches zero.
+
+  # @!macro [attach] count_down_latch_method_count
+  #
+  #   The current value of the counter.
+  #
+  #   @return [Fixnum] the current value of the counter
+
+  ###################################################################
+
+  # @!macro [new] count_down_latch_public_api
+  #
+  #   @!method initialize(count = 1)
+  #     @!macro count_down_latch_method_initialize
+  #
+  #   @!method wait(timeout = nil)
+  #     @!macro count_down_latch_method_wait
+  #
+  #   @!method count_down
+  #     @!macro count_down_latch_method_count_down
+  #
+  #   @!method count
+  #     @!macro count_down_latch_method_count
+
+  ###################################################################
+
+  # @!visibility private
+  # @!macro internal_implementation_note
+  CountDownLatchImplementation = case
+                                 when Concurrent.on_jruby?
+                                   JavaCountDownLatch
+                                 else
+                                   MutexCountDownLatch
+                                 end
+  private_constant :CountDownLatchImplementation
 
   # @!macro [attach] count_down_latch
   #
@@ -12,107 +71,7 @@ module Concurrent
   #   When the latch counter reaches zero the waiting thread is unblocked and continues
   #   with its work. A `CountDownLatch` can be used only once. Its value cannot be reset.
   #
-  # @!visibility private
-  # @!macro internal_implementation_note
-  class PureCountDownLatch < Synchronization::Object
-
-    # @!macro [attach] count_down_latch_method_initialize
-    #
-    #   Create a new `CountDownLatch` with the initial `count`.
-    #
-    #   @param [Fixnum] count the initial count
-    #
-    #   @raise [ArgumentError] if `count` is not an integer or is less than zero
-    def initialize(count = 1)
-      unless count.is_a?(Fixnum) && count >= 0
-        raise ArgumentError.new('count must be in integer greater than or equal zero')
-      end
-      super()
-      synchronize { ns_initialize count }
-    end
-
-    # @!macro [attach] count_down_latch_method_wait
-    #
-    #   Block on the latch until the counter reaches zero or until `timeout` is reached.
-    #
-    #   @param [Fixnum] timeout the number of seconds to wait for the counter or `nil`
-    #     to block indefinitely
-    #   @return [Boolean] `true` if the `count` reaches zero else false on `timeout`
-    def wait(timeout = nil)
-      synchronize { ns_wait_until(timeout) { @count == 0 } }
-    end
-
-    # @!macro [attach] count_down_latch_method_count_down
-    #
-    #   Signal the latch to decrement the counter. Will signal all blocked threads when
-    #   the `count` reaches zero.
-    def count_down
-      synchronize do
-        @count -= 1 if @count > 0
-        ns_broadcast if @count == 0
-      end
-    end
-
-    # @!macro [attach] count_down_latch_method_count
-    #
-    #   The current value of the counter.
-    #
-    #   @return [Fixnum] the current value of the counter
-    def count
-      synchronize { @count }
-    end
-
-    protected
-
-    def ns_initialize(count)
-      @count = count
-    end
-  end
-
-  if Concurrent.on_jruby?
-
-    # @!macro count_down_latch
-    # @!visibility private
-    # @!macro internal_implementation_note
-    class JavaCountDownLatch
-
-      # @!macro count_down_latch_method_initialize
-      def initialize(count = 1)
-        unless count.is_a?(Fixnum) && count >= 0
-          raise ArgumentError.new('count must be in integer greater than or equal zero')
-        end
-        @latch = java.util.concurrent.CountDownLatch.new(count)
-      end
-
-      # @!macro count_down_latch_method_wait
-      def wait(timeout = nil)
-        if timeout.nil?
-          @latch.await
-          true
-        else
-          @latch.await(1000 * timeout, java.util.concurrent.TimeUnit::MILLISECONDS)
-        end
-      end
-
-      # @!macro count_down_latch_method_count_down
-      def count_down
-        @latch.countDown
-      end
-
-      # @!macro count_down_latch_method_count
-      def count
-        @latch.getCount
-      end
-    end
-
-    # @!macro count_down_latch
-    class CountDownLatch < JavaCountDownLatch
-    end
-
-  else
-
-    # @!macro count_down_latch
-    class CountDownLatch < PureCountDownLatch
-    end
+  # @!macro count_down_latch_public_api
+  class CountDownLatch < CountDownLatchImplementation
   end
 end
